fix(clippy): green the workspace gate on Rust 1.96 / edition 2024
Regenerate common's direct-access `create.rs` from the qleany template fix
(qleany 3f96b6d6): the two `if let Replacing { if … }` blocks are now
let-chains, silencing clippy::collapsible_if. Pure syntactic change.

Fixing common — a foundational crate everything depends on — unmasked a
cascade of pre-existing clippy debt that its own failure had hidden under
`-D warnings` (a failing dependency is never linted, nor are its dependents):

- trash/binder_item work_scoping tests: drop redundant `x as u64` casts on
  values already u64 (clippy::unnecessary_cast); the real `as i64` casts stay.
- bastyde_ui/main.rs: remove the unused `app::PendingExit` import (it is used
  in backup_scheduler, not here).
- bastyde_ui/docks/inspector.rs: `inherited` is always `Some(effective_language)`,
  so compute the placeholder directly instead of `unwrap_or_default()`.
- bastyde_ui/view_models/backup_scheduler.rs: alias the per-window flush-hook
  map type (clippy::type_complexity).

`cargo clippy --workspace --all-targets -D warnings` is now clean on both the
default and `mocks` feature sets; all workspace tests still pass.

+++ crates/binder_item_management/tests/work_scoping.rs
@@ -146,9 +146,9 @@ fn merge_two_scenes_rejects_scenes_owned_by_a_different_work() {
         &mut ctx.undo,
         None,
         &MergeTwoScenesDto {
-            work_id: work_a as u64,
-            target_id: target_b as u64,
-            source_id: source_b as u64,
+            work_id: work_a,
+            target_id: target_b,
+            source_id: source_b,
         },
     );
